refactor : ChessBoardDaoImpl -> ChessDataBase 변경
 - try-with-resource로 수정
